requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
pycryptodome
python-multipart
pydantic
pytest
firebase-admin==6.4.0



